only log next-version for scripting
