+++ scripts/semantic-commits/get-binary-release-data.js
@@ -3,29 +3,11 @@ const execa = require('execa')
 const _ = require('lodash')
 const { Octokit } = require('@octokit/core')
 
+const { getCurrentReleaseData } = require('../get-current-release-data')
 const { getNextVersionForBinary } = require('../get-next-version')
 const { getLinkedIssues } = require('./get-linked-issues')
 
 const octokit = new Octokit({ auth: process.env.GITHUB_TOKEN })
-
-/**
- * Get the version, commit date and git sha of the latest tag published on npm.
- */
-const getCurrentReleaseData = async () => {
-  console.log('Get Current Release Information\n')
-  const { stdout } = await execa('npm', ['info', 'cypress', '--json'])
-  const npmInfo = JSON.parse(stdout)
-
-  const latestReleaseInfo = {
-    version: npmInfo['dist-tags'].latest,
-    commitDate: npmInfo.buildInfo.commitDate,
-    buildSha: npmInfo.buildInfo.commitSha,
-  }
-
-  console.log({ latestReleaseInfo })
-
-  return latestReleaseInfo
-}
 
 /**
  * Get the list of file names that have been added, deleted or changed since the git
@@ -117,7 +99,6 @@ const getReleaseData = async (latestReleaseInfo) => {
 
 if (require.main !== module) {
   module.exports = {
-    getCurrentReleaseData,
     getReleaseData,
   }
